# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Politics & War multiplayer strategy game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nations:
    get:
      operationId: listNations
      tags: [nations]
      summary: List all nations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NationListResponse"
  /nations/me:
    get:
      operationId: getMyNation
      tags: [nations]
      summary: Get my nation (full detail)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NationDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateMyNation
      tags: [nations]
      summary: Update my nation settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NationDetail"
  /nations/me/collect-taxes:
    post:
      operationId: collectTaxes
      tags: [nations]
      summary: Collect taxes (available every 2 hours)
      responses:
        "200":
          description: Taxes collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxCollectResult"
        "400":
          description: Too soon to collect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nations/me/change-government:
    post:
      operationId: changeGovernment
      tags: [nations]
      summary: Change government type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernmentChangeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NationDetail"
        "400":
          description: Invalid government type or cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nations/{nationId}:
    get:
      operationId: getNation
      tags: [nations]
      summary: Get a nation by ID
      parameters:
        - name: nationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── CITIES ─────────────────────────────────────────────────────────────────
...
  /nations/{nationId}/military:
    get:
      operationId: getMilitary
      tags: [military]
      summary: Get a nation's military (public info)
      parameters:
        - name: nationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilitaryPublic"
  /nations/me/military/buy:
    post:
      operationId: buyMilitary
      tags: [military]
      summary: Purchase military units
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilitaryPurchaseInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Military"
        "400":
          description: Insufficient funds or resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nations/me/military/sell:
    post:
      operationId: sellMilitary
      tags: [military]
      summary: Disband/sell military units
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilitaryPurchaseInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Military"

  # ─── WARS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NationSummary:
      type: object
      required: [id, name, leaderName, continent, governmentType, score, cityCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        leaderName:
          type: string
        continent:
          type: string
        governmentType:
          type: string
        score:
          type: number
        cityCount:
          type: integer
        flag:
          type: ["string", "null"]
        allianceId:
          type: ["integer", "null"]
        allianceName:
          type: ["string", "null"]
        beigeUntil:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    NationDetail:
      type: object
      required: [id, name, leaderName, continent, governmentType, score, cityCount, population, gdp, taxRate, taxCollectedAt, resources, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        leaderName:
          type: string
        continent:
          type: string
        governmentType:
          type: string
        score:
          type: number
        cityCount:
          type: integer
        population:
          type: integer
        gdp:
          type: number
        taxRate:
          type: number
        taxCollectedAt:
          type: ["string", "null"]
          format: date-time
        resources:
          $ref: "#/components/schemas/Resources"
        flag:
          type: ["string", "null"]
        allianceId:
          type: ["integer", "null"]
        allianceName:
          type: ["string", "null"]
        beigeUntil:
          type: ["string", "null"]
          format: date-time
        warPolicy:
          type: string
        domesticPolicy:
          type: string
        createdAt:
          type: string
          format: date-time
    NationListResponse:
      type: object
      required: [nations, total]
      properties:
        nations:
          type: array
          items:
            $ref: "#/components/schemas/NationSummary"
        total:
          type: integer
    NationUpdate:
      type: object
      properties:
        leaderName:
          type: string
        flag:
          type: string
        taxRate:
          type: number
        warPolicy:
          type: string
          enum: [Attrition, Turtle, Blitzkrieg, Fortress, Pirate, Arcane]
        domesticPolicy:
          type: string
          enum: [Manifest Destiny, Open Markets, Technological Advancement, Imperialism, Open Borders, Urbanization]
    GovernmentChangeInput:
      type: object
      required: [governmentType]
      properties:
        governmentType:
          type: string
          enum: [Democracy, Republic, Monarchy, Dictatorship, Communist, Theocracy, Oligarchy, Anarchy]
    TaxCollectResult:
      type: object
      required: [moneyCollected, nextAvailableAt]
      properties:
        moneyCollected:
          type: number
        nextAvailableAt:
          type: string
          format: date-time
        resources:
          $ref: "#/components/schemas/Resources"
    Resources:
      type: object
      required: [money, food, coal, oil, iron, bauxite, lead, uranium, gasoline, steel, munitions, aluminum]
      properties:
        money:
          type: number
        food:
          type: number
        coal:
          type: number
        oil:
          type: number
        iron:
          type: number
        bauxite:
          type: number
        lead:
          type: number
        uranium:
          type: number
        gasoline:
          type: number
        steel:
          type: number
        munitions:
          type: number
        aluminum:
          type: number

    # ─── CITIES ──────────────────────────────────────────────────────────────
...
    Military:
      type: object
      required: [nationId, soldiers, tanks, aircraft, ships, missiles, nukes, actionPointsUsed]
      properties:
        nationId:
          type: integer
        soldiers:
          type: integer
        tanks:
          type: integer
        aircraft:
          type: integer
        ships:
          type: integer
        missiles:
          type: integer
        nukes:
          type: integer
        actionPointsUsed:
          type: integer
    MilitaryPublic:
      type: object
      required: [nationId, soldiers, tanks, aircraft, ships]
      properties:
        nationId:
          type: integer
        soldiers:
          type: integer
        tanks:
          type: integer
        aircraft:
          type: integer
        ships:
          type: integer
    MilitaryPurchaseInput:
      type: object
      properties:
        soldiers:
          type: integer
          default: 0
        tanks:
          type: integer
          default: 0
        aircraft:
          type: integer
          default: 0
        ships:
          type: integer
          default: 0
        missiles:
          type: integer
          default: 0
        nukes:
          type: integer
          default: 0

    # ─── WARS ────────────────────────────────────────────────────────────────
...